requirements.txt
flask
pandas
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
azure-identity 
azure-keyvault-secrets
azure-storage-blob